Move P and TAux from NS to FP module (#13008)
