+++ modules/fluid_properties/doc/content/source/auxkernels/PressureAux.md
@@ -1,0 +1,13 @@
+# PressureAux
+
+!syntax description /AuxKernels/PressureAux
+
+Calculates the pressure given specific volume and specific internal energy.
+
+!syntax parameters /AuxKernels/PressureAux
+
+!syntax inputs /AuxKernels/PressureAux
+
+!syntax children /AuxKernels/PressureAux
+
+!bibtex bibliography
