+++ modules/fluid_properties/src/auxkernels/PressureAux.C
@@ -1,0 +1,31 @@
+#include "PressureAux.h"
+#include "SinglePhaseFluidProperties.h"
+
+registerMooseObject("FluidPropertiesApp", PressureAux);
+
+template <>
+InputParameters
+validParams<PressureAux>()
+{
+  InputParameters params = validParams<AuxKernel>();
+  params.addRequiredCoupledVar("e", "Specific internal energy");
+  params.addRequiredCoupledVar("v", "Specific volume");
+  params.addRequiredParam<UserObjectName>("fp", "The name of the user object for fluid properties");
+  params.addClassDescription(
+      "Computes pressure given specific volume and specific internal energy");
+  return params;
+}
+
+PressureAux::PressureAux(const InputParameters & parameters)
+  : AuxKernel(parameters),
+    _v(coupledValue("v")),
+    _e(coupledValue("e")),
+    _fp(getUserObject<SinglePhaseFluidProperties>("fp"))
+{
+}
+
+Real
+PressureAux::computeValue()
+{
+  return _fp.p_from_v_e(_v[_qp], _e[_qp]);
+}
