+++ modules/fluid_properties/src/auxkernels/TemperatureAux.C
@@ -1,0 +1,31 @@
+#include "TemperatureAux.h"
+#include "SinglePhaseFluidProperties.h"
+
+registerMooseObject("FluidPropertiesApp", TemperatureAux);
+
+template <>
+InputParameters
+validParams<TemperatureAux>()
+{
+  InputParameters params = validParams<AuxKernel>();
+  params.addRequiredCoupledVar("e", "Specific internal energy");
+  params.addRequiredCoupledVar("v", "Specific volume");
+  params.addRequiredParam<UserObjectName>("fp", "The name of the user object for fluid properties");
+  params.addClassDescription(
+      "Computes temperature given specific volume and specific internal energy");
+  return params;
+}
+
+TemperatureAux::TemperatureAux(const InputParameters & parameters)
+  : AuxKernel(parameters),
+    _v(coupledValue("v")),
+    _e(coupledValue("e")),
+    _fp(getUserObject<SinglePhaseFluidProperties>("fp"))
+{
+}
+
+Real
+TemperatureAux::computeValue()
+{
+  return _fp.T_from_v_e(_v[_qp], _e[_qp]);
+}
