--- modules/navier_stokes/doc/content/source/auxkernels/PressureAux.md
@@ -1,18 +1,0 @@
-<!-- MOOSE Documentation Stub: Remove this when content is added. -->
-
-# PressureAux
-
-!alert construction title=Undocumented Class
-The PressureAux has not been documented. The content contained on this page
-includes the basic documentation associated with a MooseObject; however, what is contained is
-ultimately determined by what is necessary to make the documentation clear for users.
-
-!syntax description /AuxKernels/PressureAux
-
-!syntax parameters /AuxKernels/PressureAux
-
-!syntax inputs /AuxKernels/PressureAux
-
-!syntax children /AuxKernels/PressureAux
-
-!bibtex bibliography
