--- modules/navier_stokes/doc/content/source/auxkernels/TemperatureAux.md
@@ -1,18 +1,0 @@
-<!-- MOOSE Documentation Stub: Remove this when content is added. -->
-
-# TemperatureAux
-
-!alert construction title=Undocumented Class
-The TemperatureAux has not been documented. The content contained on this page
-includes the basic documentation associated with a MooseObject; however, what is contained is
-ultimately determined by what is necessary to make the documentation clear for users.
-
-!syntax description /AuxKernels/TemperatureAux
-
-!syntax parameters /AuxKernels/TemperatureAux
-
-!syntax inputs /AuxKernels/TemperatureAux
-
-!syntax children /AuxKernels/TemperatureAux
-
-!bibtex bibliography
